requirements.txt
altair==5.4.1
langchain==0.2.16
langchain-community
nltk
pycountry
langchain_core==0.2.38
openai==1.43.1
pandas==2.2.2
pdfplumber==0.11.4
pikepdf==9.2.1
plotly==5.24.0
python-dotenv==1.0.1
streamlit==1.38.0
tqdm==4.66.5
